refactor(confirmpayment): switch if statement, use prompt via state, adjust test #1216
