+++ src/app/screens/ConfirmPayment/form.test.tsx
@@ -5,7 +5,8 @@ import { MemoryRouter } from "react-router-dom";
 import { settingsFixture as mockSettings } from "~/../tests/fixtures/settings";
 import * as SettingsContext from "~/app/context/SettingsContext";
 
-import ConfirmPayment from "./index";
+import type { Props } from "./form";
+import ConfirmPaymentForm from "./form";
 
 jest.spyOn(SettingsContext, "useSettings").mockReturnValue({
   settings: mockSettings,
@@ -22,35 +23,34 @@ jest.mock("~/common/utils/currencyConvert", () => {
 const paymentRequest =
   "lnbc250n1p3qzycupp58uc2wa29470f98wrxmy4xwuqt8cywjygf5t2cp0s376y7nwdyq3sdqhf35kw6r5de5kueeqg3jk6mccqzpgxqyz5vqsp5wfdmwtv5rmru00ajsnn3f8lzpxa4snug2tmqvc8zj8semr4kjjts9qyyssq83h74pte8nrkqs8sr2hscv5zcdmhwunwnd6xr3mskeayh96pu7ksswa6p7trknlpp6t3js4k6uytxutv5ecgcwaxz7fj4zfy5khjcjcpf66muy";
 
-// TODO: fix the test. we do not have props anymore
-// const props: Props = {
-//   origin: {
-//     location: "https://getalby.com/demo",
-//     domain: "https://getalby.com",
-//     host: "getalby.com",
-//     pathname: "/demo",
-//     name: "Alby",
-//     description: "",
-//     icon: "https://getalby.com/assets/alby-503261fa1b83c396b7ba8d927db7072d15fea5a84d387a654c5d0a2cefd44604.svg",
-//     metaData: {
-//       title: "Alby Demo",
-//       url: "https://getalby.com/demo",
-//       provider: "Alby",
-//       image:
-//         "https://getalby.com/assets/alby-503261fa1b83c396b7ba8d927db7072d15fea5a84d387a654c5d0a2cefd44604.svg",
-//       icon: "https://getalby.com/favicon.ico",
-//     },
-//     external: true,
-//   },
-//   paymentRequest,
-// };
+const props: Props = {
+  origin: {
+    location: "https://getalby.com/demo",
+    domain: "https://getalby.com",
+    host: "getalby.com",
+    pathname: "/demo",
+    name: "Alby",
+    description: "",
+    icon: "https://getalby.com/assets/alby-503261fa1b83c396b7ba8d927db7072d15fea5a84d387a654c5d0a2cefd44604.svg",
+    metaData: {
+      title: "Alby Demo",
+      url: "https://getalby.com/demo",
+      provider: "Alby",
+      image:
+        "https://getalby.com/assets/alby-503261fa1b83c396b7ba8d927db7072d15fea5a84d387a654c5d0a2cefd44604.svg",
+      icon: "https://getalby.com/favicon.ico",
+    },
+    external: true,
+  },
+  paymentRequest,
+};
 
-describe("ConfirmPayment", () => {
+describe("ConfirmPaymentForm", () => {
   test("render", async () => {
     await act(async () => {
       render(
         <MemoryRouter>
-          <ConfirmPayment />
+          <ConfirmPaymentForm {...props} />
         </MemoryRouter>
       );
     });
@@ -63,15 +63,15 @@ describe("ConfirmPayment", () => {
   test("toggles set budget, displays input with a default budget", async () => {
     const user = userEvent.setup();
 
-    await act(async () => {
-      render(
-        <MemoryRouter>
-          <ConfirmPayment />
-        </MemoryRouter>
-      );
-    });
+    render(
+      <MemoryRouter>
+        <ConfirmPaymentForm {...props} />
+      </MemoryRouter>
+    );
 
-    await user.click(screen.getByText("Remember and set a budget"));
+    await act(async () => {
+      await user.click(screen.getByText("Remember and set a budget"));
+    });
 
     const input = await screen.findByLabelText("Budget");
     const satoshis = lightningPayReq.decode(paymentRequest).satoshis || 0;
